+++ packages/elements/src/core/store/destroy-store.ts
@@ -8,6 +8,7 @@ import { log } from "../../helpers.js";
 const destroyStore = (key: string, store: Store<StoreState, StoreActions>) => {
 	store[0].stateObserver?.disconnect();
 	store[0].dispose();
+	store[0].cleanup?.();
 	Elements.stores.delete(key);
 
 	log.debug(`The store with the key "${key}" has been destroyed.`);