+++ packages/elements/CHANGELOG.md
@@ -10,5 +10,5 @@
 - Breaking change: all state values in `data-bind--` attributes must be prefixed with `$`. Ie. `data-bind--aria-label="scope:$state"`.
 - All first-party handlers now support using state as well as actions. Just prefix the value with either a `$` or `@` to denote the member type.
 - Added a new first party handler for trapping focus. This can be used via the `data-handler--trap` attribute and supports both member types.
-
+- Added a new optional cleanup callback for store modules.